Fix container-overflow in MHLO-to-LHLO
`shape.tuple_shapes(0)` returns a const-ref to the subshape, living in `shape`.

That means at some point during the assignment `shape = shape.tuple_shapes(0);`, the
vector holding the subshapes will get rewritten, while we not yet finished copying
the subshape and we end up with an out of scope read.

I fixed this by deliberately making a copy before the assignment.

PiperOrigin-RevId: 598520569