PR #8332: Store device-side time in ExecutionProfile and use it for auto-tuning
Imported from GitHub PR https://github.com/openxla/xla/pull/8332

Using ExecutionProfile for auto-tuning makes the start/stop events queued by GpuTimer wrap more tightly the kernel being auto-tuned. In particular this means that allocation/deallocation at the start/end of the executable are not counted, and the stop event is queued before the synchronisation at the end of the executable takes place.

This should improve the accuracy of the kernel execution time measurements that are used to make auto-tuning decisions.
Copybara import of the project:

--
44b718f7f69c38c8f1c8463c6b5a624558e8407b by Olli Lupton <olupton@nvidia.com>:

Store device-side time in ExecutionProfile

When auto-tuning then make sure GpuTimer tightly wraps the kernels
to be measured, and does not wrap any extra synchronization.

Merging this change closes #8332

PiperOrigin-RevId: 598810605