Validate argument minvalue of tf.random.uniform
At present the API, tf.random.uniform raises InvalidArgumentError if minval>=maxval for int dtype only. This check happens in C++ backend for int data type only.

However same is not true for float datatype. It will not raise an error and , It will consider a range ignoring min and max values in this case.

The behaviour should be same for int and float dtypes.Hence I am adding validation for same at python level itself.