Reland clean up wrong callers of FindNonTrivialHero.
We had several callers of FindNonTrivialHero which called it with a fusion
instruction. This was usually an indicator that the call was not actually needed
at all and we should just use the instruction itself as the hero.
Also adjust ChooseFusionKind to check the producer, too, whether it is a kInput
fusion.

PiperOrigin-RevId: 599125417