Creates `CalibrationComponent` that calibrates a pre-calibrated `ModuleOp`.
After calibration, statistics will be attached to the resulting `ModuleOp`. This component takes `ModuleOp` as its input and output. This implies that during the calibration process, which relies on python-level TF Session APIs (and TF runtime), it is exported to SavedModel before the calibration and imported back to ModuleOp after the calibration internally.

Tests are omitted because it requires components to be exposed to the python layer because the component depends on `PyFunctionLibrary`, which should only be injected from the python layer. Using test doubles (mocks) for `PyFunctionLibrary` has been considered but discarded, because doing so would require implementing `SaveExportedModel` properly in c++, which is an overkill for the objective of simply testing `CalibrationComponent`.

PiperOrigin-RevId: 599146078