[XLA:GPU] Consider the size of all spatial dimensions in the filter and output to detect backward filter convolutions.
Reading the comment of the original code, it mentions all dimensions, but the code only considered the first dimension. Normally, looking at one dimension is fine, but in rare cases a one filter dimension might be larger than the corrseponding output dimension, while the other filter dimension is smaller. This CL fixes the code to consider all dimensions and correctly handle the above case.

PiperOrigin-RevId: 599500724