TargetConfig: Support all fields in the StreamExecutor constructor.
The constructor that takes a StreamExecutor only sets some fields while the
GpuTargetConfigProto constructor sets all of them. Now both constructors are
in sync so the fields can be relied on.

A unit test is included to catch new fields and ensure they are set. It's rather
a change detector, but there didn't seem to be a nicer way to ensure that new
fields are properly included.

PiperOrigin-RevId: 599596417