Change the HloComputation representation for the instructions from a std::list<HloInstruction*> to a std::vector<HloInstructionInfo> (which contains an HloInstruction* and the opcode of the instruction).
Change the reverse map to use integer indices into the vector to keep
track of where instructions are located, rather than using std::list
iterator to point to the relevant instruction.

The vector is much more cache efficient than the list, and by keeping
the opcode alongside the HloInstruction pointer inside the
HloComputation::instructions() data structure, we can sometimes look
at the opcode alone to decide if the instruction is of interest
without touching the cache lines on which the HloInstruction object
lives.
PiperOrigin-RevId: 600130708