Enable priority fusion by default.
Despite extensive testing, it's possible that this will break you, since it changes XLA:GPU's fusion algorithm fundamentally. If you're affected, please set `--xla_gpu_enable_priority_fusion=false` in XLA_FLAGS and/or roll this change back. If you decide to set the flag, be aware that we're planning to remove it soon, so please create an issue, preferably with an HLO dump that shows the problem.

It is a known issue that this causes an increase in compile time. We are planning to make improvements to compile time post launch, after the current fusion logic is removed.

PiperOrigin-RevId: 600760405