Avoid confusing error messages when looking for CUDA executables
The subprocess implementation logs an error when it failed to execute
a program and doesn't allow us to gracefully handle that error.

To avoid user confusion we instead first check the existence of all the programs before we try to launch them. That's not a silver bullet
because file existence doesn't mean there exists a program we can execute,
but it should at least cut down on those error messages.

Unfortunately that also means we need to iterate over the entries in `PATH` ourselves.

PiperOrigin-RevId: 600788214