Add return type annotations to ops.py.
Primarily adds return type annotations to functions and class methods in `ops.py`
in order to improve type checking and general language services performance when
referencing types from this module.

Other type annotations, such as parameter types and variable types, were not added
in this change except as necessary to pass type checking. This change is primarily
focused on preventing the loss of type information when objects are returned from
functions and methods in this module.

PiperOrigin-RevId: 600830174