#tf-data-service Use a separate lock for `GetSnapshotSplit`.
There may be many `GetSnapshotSplit` RPCs and a few `WorkerHeartbeat`
RPCs. `GetSnapshotSplit` RPCs are slower than `WorkerHeartbeat`. If
they share a mutex, `WorkerHeartbeat` RPCs could be slowed down.

Also fix a race condition when restoring repetition indexes. If the
snapshot manager creates repetition directory N and then dies, the
restored repetition index should be N-1 and the split provider is in
the end of split state. Otherwise, when the worker requests the next
split for repetition N, it may return end of sequence immediately,
which increases the worker's repetition index to N+1. Consequently,
one repetition is lost. The fix is increasing the repetition index
when resetting the split provider, not by counting the directories.

PiperOrigin-RevId: 600880725