Do not match `RewriteQuantizedDotGeneralOpToTflFullyConnectedOrBatchMatmulOp` when dot_general is not followed by a requantization.
The case without a following requantization is partially supported by the other pattern, `RewriteUpstreamQuantizedDotGeneralOpToTflFullyConnectedOp`, but this change guards against an edge case where the output is an i32 quantized tensor.
Without this change, the program will crash when a `dot_general` produces an `i32` quantized tensor but is not followed by a requantization.

PiperOrigin-RevId: 601407496