NFC: Make the tiling scheme batch dimension not special.
For some reason the batch dimension is handled separately, which results in
a few annoying things:
- repetition of bounds computation
- tile sizes have only two elements, whereas everything else has 3. Of course
  the indices don't match up.

The EmitTileRec is a bit odd in some places, this is because I was trying to
minimize the IR changes.

I had to make a few changes to the reduction emitter. It had its own copy of
the x offset computation, which was used in weird ways (to compute the y start?).

I didn't regenerate the hlo test checks, to optimize for reviewability. Therefore,
some value numbers are now out of order.

The next step is to either get rid of vector size or the two different X indexing
modes. Eventually both, hopefully.

PiperOrigin-RevId: 601736735