Speed up XlaComputation post order computation by using bitvector for DFS.
Now that HloInstructions are stored in a vector in their parent
HloComputation, we have a convenient dense index for each instruction.
This is present in the instruction_indices_ hash table in
HloComputation.  This change also stores it in the HloInstruction
itself.  This allows ready access during the DFS that is used to
compute various post orders.

Once we have a dense index easily available for every node that we
visit, we can optimize the DFS in a few ways:

(1) We can use two bits to keep the traversal state for each node in a
dense bit array (in a new implementation of the VisitMap abstraction,
which previously used a more complex flat array or hash table
implementation, depending on the number of nodes to be traversed).

(2) We can cheaply get the state of each outgoing node we are
considering during the DFS.  If the state of the node is already
kVisited, we can avoid pushing it onto the stack at all.  At the cost
of looking up the state in the VisitMap, we avoid a push and a pop
onto the DFS stack and a loop iteration in the DFS.

Good for about a 0.6% to 4.73% on some representative large model compilation
benchmarks.

PiperOrigin-RevId: 602515283