Fix `FuseFullyConnectedAndMul` to check if FullyConnected has only one use, that is the LHS of Mul Op. Otherwise this will duplicate the fullyconnected op to serve the remaining uses.
For example, consider this MLIR graph with the output of FC used by multiple ops:
```
  %0 = "tfl.fully_connected"(%arg0, %cst0, %cst1) {fused_activation_function = "NONE", keep_num_dims = false, weights_format = "DEFAULT"} : (tensor<4x2xf32>, tensor<2x2xf32>, tensor<2xf32>) -> tensor<4x2xf32>
  %1 = "tfl.reshape"(%0, %cst3) : (tensor<4x2xf32>, tensor<2xi32>) -> tensor<1x8xf32>
  %2 = "tfl.mul"(%0, %cst2) {fused_activation_function = "RELU6"} : (tensor<4x2xf32>, tensor<2xf32>) -> tensor<4x2xf32>
```

The expected output for the fusion is a `NEW` FC op with the Mul op fused into the FC kernel, and the `OLD` FC op deleted. But the reshape-op above does not need/expect a fc with mul fused. So, the 'OLD' fc op is still around to feed into the ReshapeOp. Hence we are computing two fully-connected ops in the graph instead of one.

This particular bug has always been there but got triggered on this customer model(b/318735772) by this `recent` change cl/553508906.

This is a broader issue that needs to be fixed for all the BinaryOp(*FC_Op*) fusion optimization variants on FullyConnected Op, Convolutions, etc.

PiperOrigin-RevId: 602536477