Restructure NVPTXCompiler::ChooseLinkingMethod
This is separating the caching mechanism in NVPTXCompiler::ChooseLinkingMethod from
the actual linking method choosing logic.

This allows for standard return-early style and make the logic easier to follow (less nesting).

I'm doing this now because I want to add another type of linking in a subsequent CL/PR.

I'm also making ChooseLinkingMethod take a reference to DebugOptions because I will need that in the next CL/PR.

PiperOrigin-RevId: 602954817