[XLA:GPU] Fix tolerance issue in TritonGemmTest.SplitLhsNoncontractingTransposeRhs.
The test is flaky due to numerical imprecisions frequently significantly
exceeding the previous thresholds. We now use booleans as parameters to
restrict the domain of the inputs to the `dot` operation. The result can
now be represented with 100% accuracy using f16, meaning that the test can
no longer fail flakily.

PiperOrigin-RevId: 602976585