[XLA:GPU] Add proper checks for code duplication to priority_fusion.
We were using fusion_node_indexing_evaluation which just supports checks
whether a single producer can be fused into a fusion, but we called it also
with producer fusions which were then not properly evaluated. We already have a
different way to check this via cost_analysis which is also used in
FusionMerger, so we should be using that instead.

PiperOrigin-RevId: 603267214