Add `--build_tests_only` to build-only env
We need this flag to restrict building only "_test" and "test_suite" rules in the build-only continuous job. While this flag is defined in TF's root .bazelrc, it is not picked up because it is defined as a "test:" config. We do not want to add it as a "build:" config in the .bazelrc because then we cannot build non-test targets with RBE cross-compilation.

PiperOrigin-RevId: 603756845