Add `tensorflow-macos` as a shim package
Before Google had internal hardware support for Apple Silicon, macOS Arm64 wheels used to be built on third-party CI and were uploaded to `tensorflow-macos`. In order to not break users who use `pip install tensorflow`, we used to upload a small "installer" wheel to `tensorflow`. This installer wheel would automatically download and install the correct `tensorflow-macos` package.

Starting with TF 2.16, Apple Silicon packages are going to be uploaded directly to the "tensorflow" project on PyPI. In order to not break users who are still using `tensorflow-macos`, we upload the installer wheel instead to "tensorflow-macos" and add "tensorflow" as its dependency. Please note that this will go away in TF 2.17 and `tensorflow-macos` will be considered deprecated.

PiperOrigin-RevId: 603757844