Comments and presentation of WIP Windows support for CI
As far as I can tell, this encapsulates what I had in-progress while I
was working on getting Windows builds to work better in Docker containers.
I had a Windows Kokoro machine running Docker that I used to test this,
and was able to get builds working, but wasn't feeling certain about the
configuration, and hadn't integrated anything with the CI system yet.
I've included a lot of comments explaining what I was working on.

One detail that's left out of the comments is that I'm not sure how
the recent changes to the Pip package build process factor in to this.