[XLA:GPU] Add check that computation is found in TritonFilecheckTest.
TritonFileCheckTest::CreateTritonIrAndFilecheck fails with an obscure segmentation fault in TritonFusionAnalysis::Execute if an incorrect name for the triton function is passed in. This makes it seem like the issue is around tiling analysis when it is just a test typo. This change adds a ret_check to clarify that the failure is from the call to GetComputationWithName.
PiperOrigin-RevId: 605007087