Remove public transitive includes from stream_executor.h
`stream_executor.h` is transitively including a bunch of headers that are actually public headers and should be included directly by users.

In this change I'm removing all of them except:
- `multi_platform_header.h`
- `stream.h`

Removing `multi_platform_header.h` requires a few changes in user code which I prefer to put in a separate CL.

Removing `stream.h` is more tricky. To make it compile I have to add more than a hundred includes in different places - often in files that don't even use the `Stream` class directly because cyclic dependencies prevent me from adding the include where it's supposed to be. Therefore before doing that change we have to discuss how we envision the dependency structure of stream_executor.

PiperOrigin-RevId: 605228718