#tf-data Serialize tensors separately in tf.data iterator save/restore.
Previously, the entire checkpoint would be serialized as a single proto, introducing a 2GB size limit. Now each slice of the checkpoint will be serialized separately, matching how tf.data iterators are saved in tf.train.CheckpointManager.

This breaks backwards compatibility of previously-written checkpoints, but we only guarantee checkpoint restorability when the same CL is used for both writing and reading.

PiperOrigin-RevId: 605681289