Fix calculation in resize_bicubic_op_test
Ensure that the baseline calculation matches the optimised
case even when on a compiler and architecture that can use
higher precision intermediate values which result in different
calculated values. This is done by splitting one calculation
into two parts so that the intermediate value is truncated
so that it can be stored in a float. This ensures that the
baseline and the optimised calculations do not diverge.