Implement minimal selective quantization feature supporting denylisting.
This change provides a minimal config and implementation for denylisting quantizable units (== lifted functions).
With this change, users will be a able to denlylist quantization for specific quantizable units by specifying the config as;

```textpb
specs [
  {
    matcher { function_name { regex: "composite_dot_general.*" } }
    method { no_quantization{} }
  },
]
```

This change also includes an implementation of `TestLiftQuantizableSpotsAsFunctionsWithQuantizationSpecsPass`, which is `LiftQuantizableSpotsAsFunctionsPass` with predefined `QuantizationSpecs`.

PiperOrigin-RevId: 607221399