PR #9528: [cudnn] Use new cudnn frontend API for flash attention
Imported from GitHub PR https://github.com/openxla/xla/pull/9528

This patch adds support for new cudnn frontend API for flash attention. Additional tests are not required, as this is a replacement for the previous implementation.
Copybara import of the project:

--
30545115d703dbdf396af00a3246e32d35d4abe7 by Shraiysh Vaishay <svaishay@nvidia.com>:

[cudnn] Use new cudnn frontend API for flash attention

This patch adds support for new cudnn frontend API for flash attention.
Additional tests are not required, as this is a replacement for the
previous implementation.

Merging this change closes #9528

PiperOrigin-RevId: 607234041