[XLA] Add a legalization pass for host offloader
Things that change the code in such a way that is compatible for host_offloader.
As of now move an legalization step that splits broadcasts in it that was already in host_offloader
and add a new one that moves layout changing copies out of the path of the data living on the host.
(moving it before the it gets copied on the host).

PiperOrigin-RevId: 607750059