Make sharding identification symmetric for aliases.
Refactors and changes the behavior of the sharding identification pass as follows:
1. We first try to determine the sharding for each input/output
2. If we couldn't find sharding for some inputs/outputs which are aliased, try to see if we were able to find the sharding for their corresponding outputs/inputs and use those respectively.
3. If there are still some inputs/outputs without sharding, default to replicated or maximal sharding.

This makes the sharding determination from aliases symmetric for inputs and outputs. Prior to this change, the logic to determine output sharding would short-circuit when an input sharding was found for the aliased output. This is a problem because: if the input sharding had already defaulted to replicated/maximal sharding, the corresponding aliased output would also default to it even though it was possible to determine a more accurate sharding configuration for that particular output independently.

PiperOrigin-RevId: 608209734