Fix the behaviour of the if_gpu_is_configured macro
If the Bazel build is configured for the first time on a remote worker pool like RBE, then ROCm configure script won't have access to the true value of the `TF_NEED_CUDA` environment variable and therefore won't return true from `if_gpu_is_configured` if CUDA is available, but not ROCm.

The fix is to explicitly make `TF_NEED_CUDA` part of the remote execution environment in which the ROCm configure script runs.

PiperOrigin-RevId: 608739095