Allow VHLO to interoperate with TFLite dialect.
This is accomplished by wrapping StableHLO sections in UnrealizedCastConversionOp and converting to VHLO to do version management. Selectively converting ops and materializing unrealized conversions is accomplished using MLIR Dialect Legalization.

The test case ensures both that the StableHLO->VHLO generation is as expected, as well as that a round trip back to StableHLO is lossless and matches an identity parse of the test file.

PiperOrigin-RevId: 608779149