[Distributed Eager] Add `TF_EAGER_REGISTER_ABSTRACT_FUNCTIONS_LOCAL_ONLY` environment variable.
The default behavior of the distributed eager runtime is to register all functions with all workers. In some cases, this is unnecessary, because the `ProcessFunctionLibraryRuntime` will separately send post-optimized versions of the relevant functions when running a multi-device function across workers; and it can also be take a long time when running large models on large clusters.

This change adds a new environment variable to control this behavior. Setting `TF_EAGER_REGISTER_ABSTRACT_FUNCTIONS_LOCAL_ONLY=1` will avoid registering functions remotely.

PiperOrigin-RevId: 609090780