Do not crash on ptxas failures
This leads to better UX for Pallas:GPU users e.g. when the underlying GPU
does not support a feature used by the Triton compiler.

This change also fixes a caching bug in NVPTXCompiler::CompileGpuAsmOrGetCachedResult,
which previously incorrectly cached an empty cubin if the compilation failed,
instead of consistently failing on each compilation attempt, if the first
one failed.

PiperOrigin-RevId: 609123602