[xla:gpu] Convert StreamExecutorInterface Memcpy (host src) to return absl::Status instead of bool.
This changes the return type of StreamExecutorInterface's virtual Memcpy function (taking a host src to device dest) from a bool return value to return absl::Status instead, which allows for more localized failures. It changes all overrides to return absl::Status, but for now does not change the return type of calling functions.

PiperOrigin-RevId: 609539163