Define BUILD_NUM_JOBS value only if not set by environment
BUILD_NUM_JOBS is not set by default, so unless it is set by the environment, it leads to a boundless number of processes being started upon compilation. The previous patch set this value, but prevented override by the environement. This commit sets the value, but it allows this value to be overridden by the environment if the users explicitely decides to.