Fix handling of non-finite determinants.
The previous routine would fail if the LU decomposition had a 0 on the diagonal,
or if it contained infs or nans.  We had a "hack" to try to fix the zero case,
but this would cause NaN determinants to return zero.

For a more robust sign computation, use the `.sign()` routine rather than divisions.
This eliminates the need for special handling in case `abs_diag` contains a zero,
which fixes the other `inf`/`nan` cases.

Fixes #63052.

PiperOrigin-RevId: 610860994