Skips wait for remote tensor handle to be ready based on an environment variable
This wait is needed when both a tensor handle (input to a function) and a component function can be remote to ensure that the tensor handle is ready before the function executes. In cases, where it can be guaranteed that the function will never be sent before the input to the function, this wait can be skipped.

PiperOrigin-RevId: 612221424