[FunctionLibrary] Add lazily evaluated `creator` overload for `FunctionLibraryDefinition::AddOptimizedFunctionGraph()`.
In common use, the graph stored under this API is loaded from some external source and only accessed once (upon initial function instantiation). Because optimized function graphs can be multiple gigabytes in size, it wastes RAM to hold onto this state for the entire program execution. Therefore, the new overload allows callers to provide a `creator` callback that (e.g.) loads the graph from external storage, and then free the graph once it is no longer used.

In addition, `FindOptimizedFunctionGraph()` now returns an owned copy of the graph, so the `ProcessFunctionLibraryRuntime` can use move constructors to reduce the amount of work and avoid keeping two copies of the graph structure resident in RAM during instantiation.

PiperOrigin-RevId: 612304521