[AbstractFunction] Use `FunctionRecord` instead of `FunctionDef` to enable sharing.
Previously, the `Runtime::RegisterConcreteFunction()` method would create two copies of each registered `FunctionDef`: (1) in the `AbstractFunction` returned by the call, and (2) in the `AbstractContext` with which the function was registered (often an `EagerContext`). By wrapping that function in a `FunctionRecord` we can share the immutable function message in both places.

PiperOrigin-RevId: 612525693