[PFLR] Add `TF_PFLR_PARALLEL_INSTANTIATE_THRESHOLD` to control parallel instantiation.
Currently, when there are any more than 8 partitions in a multi-device function, TensorFlow will attempt to instantiate those partitions in parallel. However, in some cases, a very large number of partitions will lead to running out of host memory, because the underlying protobufs are copied several times.

If your job is OOMing during startup, as a workaround, set `TF_PFLR_PARALLEL_INSTANTIATE_THRESHOLD` to a value larger than the total number of devices in the job.

In addition, this change frees the subgraph-that-is-converted-to-`FunctionDef` sooner in the instantiation flow (and crucially before attempting to send the RPC registering the function), which should reduce the peak memory consumption even without limiting parallelism.

PiperOrigin-RevId: 612528479