Move libnvptxcompiler support into asm_compiler.cc
Currently we use the ptxas command line tool to lower from
PTX to SASS and I'm working on adding a second way using
the libnvptxcompiler library. This way is currently behind
a flag and the NVPTXCompiler checks the flag and calls
into the different lowering options.

This is a problem for some XLA consumers because there
are quite a few which call the function `CompileGpuAsm`
from asm_compiler.cc directly in the code. This function
only supports ptxas compilation.

To also support these users I'm moving the branching
into `CompileGpuAsm`.

1. I rename the original `CompileGpuAsm` to `CompileGpuAsmUsingPtxAs`
2. I create a new function `CompileGpuAsm` which checks
the debug options and then either calls `CompileGpuAsmUsingPtxAs`
or `CompileGpuAsmUsingLibNvPtxCompiler`.

PiperOrigin-RevId: 612758451