Remove potentially buggy code.
Passing two times result_types to mapConvertOpToStdScalarOp is incorrect if we
are trying to map from an unsigned type to a signless type. Instead, for the
Convert op, the mapConvertOpToStdScalarOp function should be called directly
which allows to specify different arguments for targetTypes and resultTypes
parameters. targetTypes will be used to decide whether arith.fptoui or
arith.fptosi is used, and resultTypes is used to select the signless type that
is compatible with those arith ops.

PiperOrigin-RevId: 612800873