[XLA:GPU/CPU] Lower bfloat16 to the LLVM type instead of emulating it through i16
This does away with a couple of hacks and gives LLVM more room for
optimization. It's also necessary to make use of more bf16 hardware
instructions, such as in Hopper GPUs. Don't remove the upcasting to f32 on the
HLO level yet as that has numerical implications.

Relanding after fixing bfloat on AArch64 in LLVM.

PiperOrigin-RevId: 612905678