[PFLR] When instantiating functions serially, block on component function instantiation.
Previously, we instantiated functions sequentially when the number of subgraphs is less than a threshold, and made the threshold configurable to control memory usage. However, the implementation ends up invoking the `InstantiateRemote()` method asynchronously, which can lead to a large number of serialized buffers being resident in RAM at once. This change serializes the entire instantiation, which further reduces memory pressure.

PiperOrigin-RevId: 612957391