[PJRT] Update PjRtBuffer::layout() to return PjRtLayout.
Also updates callers (dlpack and similar code in py_array.cc). This
further breaks the GPU plugin because I haven't updated the C API to
use something similar to PjRtLayout. I will do so in a future series
of changes.

I went back and forth on whether layout() should return by value or
reference (unique_ptr or const ref). I went with value to make the
memory management easier -- it's hard to copy a PjRtLayout& without
access to the subclass definition, and at least in Python we'll want
layout objects to potentially outlive the buffers or executables they
came from.

PiperOrigin-RevId: 612971532