Fix data race in HloInstruction::BackendConfigRep
`xla::HloInstruction::BackendConfigRep::GetRawString` gets called from multiple threads at the same time in Triton Autotuning.

The implementation of this function caches its results in a mutable member variables in the object.
But this mutable member is not protected by a mutex which leads to concurrent write accesses from
multiple threads and therefore crashes or memory leaks.

So this change is adding a mutex to `mutable std::string raw_string_`.

PiperOrigin-RevId: 613618629