Rewrite and simplify epilogue handling.
The old approach ("isolated and injected") is confusing and doesn't work.
The new approach is to create a separate subgraph for the epilogue and
emit that as its own function. That way, only the epilogue needs to deal
with injected values.

This fixes most of the miscompiles in JAX. Transpose doesn't handle non-
trivial epilogues yet, but that should be an easy fix, similar to what's
done with reduce.

PiperOrigin-RevId: 616090173