Add new overload of `GraphToFunctionDef()` that can consume the graph.
Previously, we copied all of the node data from the input subgraph to the `FunctionDef`. This change reduces the memory footprint of multi-device function instantiation by (where possible) `std::move`-ing the bulk of the node data from the input graph to newly created `NodeDef` members of the resulting `FunctionDef` protocol buffer.

PiperOrigin-RevId: 617641191