Remove the `OpSet` alias in `mlir::quant` namespace.
It is defined in the `tensorflow::quantization::` namespace (as part of `QuantizationOptions` proto), but has been unexpectedly exposed under `mlir::quant` due to a `using` directive in `attrs_and_constraints.h`. Remove this exposure and use the one in the `tensorflow::quantization::`.

This change also fixes some include warnings along the way.

PiperOrigin-RevId: 618067952