Do not fuse gelu into FP8 cublasLT matmul on CUDA versions less than 12.4.
This causes cublas LT to give an error in certain cases. Specially, it always causes an error on Ada and causes an error when fast accumulation is disabled on Hopper, at least in the cases I've tried. I am just disabling it in general for older CUDA versions in case there are other scenarios where it fails.

The issue actually only occurs on CUDA versions less than 12.3.2, but there is no good way of checking the patch version of CUDA, so I'm disabling it on CUDA versions less than 12.4. There is a CUBLAS_VER_PATCH with the patch version of cublas but this is undocumented so I'd rather not use it.

Thank you @philipphack and @artem-b for helping me debug this.

PiperOrigin-RevId: 618135629