Use indexing map symbol rescaling in the ReduceWindow emitter
We current codegen the reduce window op as a loop nest with the inner loop
iterating over the window and accumulating input values.

If the op had a base dilation set we would still generate the same inner
loop but the loop body now woudl have an additional condition that checks
whether we are in bounds of the dilation.

This change makes use of IndexingMap's symbol rescaling which results
in the generation of an inner loop with fewer iterations. It also avoids
the in-bounds check by only iterating over the tensor elements that
actually need to be accumulated.

PiperOrigin-RevId: 619893695