Support lifting of same shape bias for `stablehlo.convolution`.
Although it is common to write models where 1D constant would be broadcasted, it is possible to explicitly give bias with the desired shape.
The examples in `odml_coverage_test` are such.
Handle such cases where the bias has the same shape as target accumulation.

Additionally, add `FindOperandType` for finding an operand of specific type and respective tests.

PiperOrigin-RevId: 621749920