presubmits: Add a presubmit for CHECK and related macros.
Generally it's a better idea to find a way to report the error rather than to
crash the process. When XLA is used in a server, crashes expose the server to
potential queries of death when unexpected data or API calls occur.

Removing existing crashes is a longer-term fixit-type effort, but at least we
can prevent new intentional crashes from slipping in.

PiperOrigin-RevId: 621952801