[IFRT] Add fast pointer equality test for `DeviceList` internal state
Copied `DeviceList`s share their internal state with the source `DeviceList`.
We leverage it to make comparison faster between two `DeviceList` that are a
(transitive) copy of the other.

The single-device `DeviceList` does not use pointer equality test because the
internal state is not wrapped in a shared pointer, but this case will not
benefit from pointer equality test anyway.

PiperOrigin-RevId: 622347506