#tf-data Implement index_flat_map dataset.
This is similar to flat_map, but supports global shuffling.

In addition to a `map_func`, the user needs to provide an `index_map_fn`.
Given an index in the flattened dataset, the `index_map_fn` returns a
(element index, offset) tuple that represents the index of the element in the
unflattened dataset, and the offset in the unflattened element.

PiperOrigin-RevId: 622952111