Downgrade rendezvous abortion logging.
This produces a confusing warning when iterating over a `tf.data.Dataset`.

For example, see the current state of https://www.tensorflow.org/tutorials/keras/text_classification#load_the_dataset (sorry for the nonfutureproof link; I can't figure out how to link to a specific version of this tutorial that includes the output), where this warning shows up in the output of the iteration over `raw_train_ds.take(1)`.

PiperOrigin-RevId: 623300763