Temporarily disable newer cuDNN fMHA features due to miscompile
I'm trying to make XLA compatible with cuDNN 9 and stumbled upon
a miscompile in the fMHA rewriter.

It seems to be related with the cause mask pattern matcher, at least
forcing the causal mask flag to true, makes the broken test path.

I extract a reproducer and added an (integration) test which
can be removed or converted into a proper unit test once this is fixed.

IN the meantime I disable all fMHA features that require a version
of cuDNN 8.9.6+.

PiperOrigin-RevId: 623470743