[XLA:GPU] Refactor HloFusionAdaptor to a flat structure.
There is a need for HloInstructionAdaptor to know it's parent HloFusionAdaptor to be able to distinguish fusion instruction for operands and users. Current nested structure of HloFusionAdaptor prevents us from having a convenient single parent pointer.

This CL add `parent` field to HloInstructionAdaptor, but doesn't set or use it consistently yet.

PiperOrigin-RevId: 626307316