PR #11611: Define GOOGLE_CUDA to turn on #11568
Imported from GitHub PR https://github.com/openxla/xla/pull/11611

The new NVTX annotations added in #11568 will not appear unless GOOGLE_CUDA is defined. The part of #9896 that was supposed to avoid this was rolled back in c4b031bf9522685f998ecbd12570f29b391495ec.

cc: @cheshire
Copybara import of the project:

--
80e8f28422aa1accfd4b972839b5f8ae1b7ff74b by Olli Lupton <olupton@nvidia.com>:

Define GOOGLE_CUDA to turn on #11568

The new NVTX annotations added in #11568 will not appear unless
GOOGLE_CUDA is defined. The part of #9896 that was supposed to avoid
this was rolled back in c4b031bf9522685f998ecbd12570f29b391495ec.

Merging this change closes #11611

PiperOrigin-RevId: 626959591