[XLA:GPU] Avoid segfault in serializing autotune results.
In some cases, the debug options reference may become invalid in
GpuCompiler::RunHloPasses, at which point serializing segfaults upon access to
the debug options.

The reason for the invalid reference is not clear, as all methods operate on
unique_ptr::get(), not on raw pointers that would invalidate it, but taking a
copy of a modest-sized variable is not the end of the world from a performance
perspective even if it's not ideal.

A never-used call to GetAutotuneConfig is also removed from RunHloPasses.

PiperOrigin-RevId: 627461704