Adds default move constructors to `tensorflow::XlaCompiledFunction`.
This class already deletes the copy constructors, but does not enable the move constructors. As written today, users of `tensorflow::XlaCompiledFunction` must wrap their instances in a `std::unique_ptr` (or similar) in order to move this instance. That solution would add indirection and proliferate null checks that is not needed after this chance.

PiperOrigin-RevId: 627474665