Accounts for the number of buffer elements in the memory usage of the model
Previously, when autotune optimizes the model, it did not accounts for the usage of the actual number of buffered elements. This causes the memory usage to fluctuate and potentially makes tf.data to use RAM more than the ram budget tf.data supposed to respect.

After this change, tf.data's memory usage is expected to be more stable and less likely to consume more than the ram budget.

PiperOrigin-RevId: 627926221