PR #11592: Add a TritonFusionAnalysis util to query an op's scope.
Imported from GitHub PR https://github.com/openxla/xla/pull/11592

Add a TritonFusionAnalysis util to query an op's scope.
This will be used in cuDNN compiler for deciding an op's dimensions in the followup PR.
Copybara import of the project:

--
33d4a59ebf970689a4c05c1847152c9a6aab632f by Elfie Guo <elfieg@nvidia.com>:

Add a TritonFusionAnalysis util to query an op's scope.

Merging this change closes #11592

PiperOrigin-RevId: 628015848